requirements.txt
Pillow==9.3.0
Pillow==9.5.0
Requests==2.29.0
streamlit==1.21.0
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
~treamlit==1.15.0
